+++ src/generators/finance/bank.ts
@@ -1,6 +1,11 @@
 import { fromFormat } from '../util/string';
 import { getLocale } from '../../configuration';
 
+/**
+ * @category Generators
+ * @subcategory Finance
+ */
+
 // https://www.iban.com/structure
 // https://www.investopedia.com/terms/i/iban.asp
 export function IBAN(): string {