+++ src/generators/internet/address.ts
@@ -6,7 +6,7 @@ import { getLocale } from '../../configuration';
 
 /**
  * @category Generators
- * @subcategory Geographic
+ * @subcategory Internet
  */
 
 interface IAddress {