+++ src/generators/internet/index.ts
@@ -1,0 +1,7 @@
+import { authentication } from './authentication';
+import { password } from './password';
+
+export const Internet = {
+	authentication,
+	password
+};